removed some useless files, restructred project
