+++ frontend/public/index.html
@@ -27,7 +27,7 @@
       work correctly both with client-side routing and a non-root public URL.
       Learn how to configure a non-root public URL by running `npm run build`.
     -->
-    <title>React App</title>
+    <title>Americorps</title>
   </head>
   <body style = "background-color: #07192c;">
     <noscript>You need to enable JavaScript to run this app.</noscript>
