--- frontend/src/App.css
@@ -1,16 +1,0 @@
-
-.App {
-  background-color: transparent;
-}
-
-.App-header {
-  
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: left;
-  font-size: calc(60px + 2vmin);
-  color: white;
-}
-
-
