--- frontend/src/components/AshStyle.css
@@ -1,174 +1,0 @@
-/* RICK AND ROSIE IMAGE */
-.HomeImage{
-    position:absolute; 
-    width: 1024px;
-    height: 524px;
-    top:171px;
-    left:73px;
-    mix-blend-mode: normal;
-    margin: 0;
-    padding: 0;
-  }
-  
-  /* WELCOME TO AMERICORPS*/
-  .AmericorpsTitle{
-    margin: 0;
-    padding: 0;
-    position: absolute;
-    width: 899px;
-    height: 50px;
-    left: 73px;
-    top: 800px;
-  
-    font-family: 'Lato', sans-serif;
-    font-style: normal;
-    font-weight: 800;
-    font-size: 60px;
-    line-height: 72px;
-    display: flex;
-    align-items: center;
-  
-    color:white;
-  
-    box-sizing: border-box;
-  
-  }
-  
-  /* MONTH YEAR */
-  .UsersDate{
-  position: absolute;
-  width: 244px;
-  height: 72px;
-  left: 37px;
-  top: 150px;
-  
-  font-family: Roboto;
-  font-style: normal;
-  font-weight: 200;
-  font-size: 36px;
-  line-height: 42px;
-  display: flex;
-  align-items: center;
-  text-align: center;
-  
-  color: #FFFFFF;
-  }
-  .fillerText{
-    position:absolute;
-    width: 840px;
-    height: 1800px;
-    left: 73px;
-    font-style: normal;
-    font-weight: 800; 
-    line-height: 36px;
-    display: flex;
-    align-items: center;  
-    font-size: 30px;
-    font-family: 'Lato', sans-serif;
-    color:white;
-  
-  }
-  
-  /* [ + ] Add User */
-  .AddUserButton{
-  position: absolute;
-  padding:10px;
-  width: 180px;
-  height: 51px;
-  left: 1266px;
-  top: 156px;
-  
-  font-family: Roboto;
-  font-style: normal;
-  font-weight: 200;
-  font-size: 24px;
-  line-height: 28px;
-  display: flex;
-  align-items: center;
-  text-align: center;
-  background: none;
-  border: none;
-  color: #FFFFFF;
-  }
-  
-  /* [ / ] Edit User */
-  .EditUserButton{
-    position: absolute;
-    width: 50px;
-    height: 60px;
-    left: 1200px;
-    top: 315px;
-    font-style: normal;
-    font-weight: 900;
-    font-size: 24px;
-    line-height: 28px;
-    display: flex;
-    align-items: center;
-    text-align: center;
-    background: none;
-    border: none;
-    color: #FFFFFF;
-  
-  }
-  
-  /* [ 8 ] DELETE USER */
-  .DeleteUserButton{
-    position: absolute;
-    width: 50px;
-    height: 60px;
-    left: 1270px;
-    top: 315px;
-    font-style: normal;
-    font-weight: 900;
-    font-size: 24px;
-    line-height: 28px;
-    display: flex;
-    align-items: center;
-    text-align: center;
-    background: none;
-    border: none;
-    color: #FFFFFF;
-  }
-  
-  .spaceText{
-    padding:10px
-  }
-  
-  /* Total Monthly Hours*/
-  .MonthlyTotal{
-    position: absolute;
-    width: 320px;
-    height: 47px;
-    left: 63px;
-    top: 958px;
-    
-    font-family: Roboto;
-    font-style: normal;
-    font-weight: 200;
-    font-size: 24px;
-    line-height: 28px;
-    display: flex;
-    align-items: center;
-    
-    color: #FFFFFF;
-  }
-  
-  /* Total Yearly Hours*/
-  .YearlyTotal{
-    position: absolute;
-    width: 296px;
-    height: 47px;
-    left: 1097px;
-    top: 958px;
-    
-    font-family: Roboto;
-    font-style: normal;
-    font-weight: 200;
-    font-size: 24px;
-    line-height: 28px;
-    display: flex;
-    align-items: center;
-    text-align: right;
-    
-    color: #FFFFFF;
-  }
