+++ frontend/src/components/Header/index.module.css
@@ -1,6 +1,6 @@
 .header{
     color: aliceblue;
-    background-color: #0d213c;
+    background-color: transparent;
     padding: .5% 40% .5% 40%;
     justify-content: center;
     margin: none;
