+++ frontend/src/components/HomeBody/index.module.css
@@ -53,7 +53,7 @@
     font-size: 100%;
 }
 .buttonUserClick{
-    background-color: #0d213c;
+    background-color: #212f42;
     border:none;
     color:#b5b5b5;
     padding: 2% 15%;
@@ -63,7 +63,7 @@
     font-size: 100%;
 }
 .buttonAdminClick{
-    background-color: #0d213c;
+    background-color: #212f42;
     border:none;
     color:#b5b5b5;
     padding: 2% 15%;
