--- frontend/src/components/Navbar.module.css
@@ -1,135 +1,0 @@
-.Nav {
-    height: 100px;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    font-size: 1.2rem;
-    position: sticky;
-    top: 0;
-    z-index: 999;
-    border-bottom: 1px solid #fff;
-}
-
-.NavContainer {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    height: 100px;
-    max-width: 1500px;
-}
-
-.NavLogo {
-    color: #fff;
-    justify-self: start;
-    margin-left: 20px;
-    cursor: pointer;
-    text-decoration: none;
-    font-size: 2rem;
-    display: flex;
-    align-items: center;
-}
-
-.LogoImg {
-    width: 75px;
-    max-width: 100%;
-    height: 75px;
-    max-height: 100%;
-    object-fit: cover;
-    padding-right: 20px;
-}
-
-.Menu {
-    display: grid;
-    grid-template-columns: repeat(4, auto);
-    grid-gap: 10px;
-    list-style: none;
-    text-align: center;
-    width: 60vw;
-    justify-content: end;
-    margin-right: 2rem;
-}
-
-.MenuItem {
-    height: 100px;
-}
-
-.MenuItemLink {
-    color: #fff;
-    display: flex;
-    align-items: center;
-    text-decoration: none;
-    padding: 0.5rem 1rem;
-    height: 100%;
-}
-
-.MenuItemLink:hover {
-    border-bottom: 6px solid #fff;
-    transition: all 0.2s ease-out;
-}
-
-.fa-bars {
-    color: #fff;
-}
-
-.MenuIcon {
-    display: none;
-}
-
-@media screen and (max-width: 960px) {
-    .Menu {
-        display: flex;
-        flex-direction: column;
-        width: 100%;
-        height: 90vh;
-        position: absolute;
-        top: 80px;
-        left: -100%;
-        opacity: 1;
-        transition: all 0.5s ease;
-    }
-
-    .Menu.active {
-        background: #242222;
-        left: 0;
-        opacity: 1;
-        transition: all 0.5s ease;
-        z-index: 1;
-    }
-
-    .MenuItemLink {
-        text-align: center;
-        padding: 2rem;
-        width: 100%;
-        display: table;
-    }
-
-    .MenuItemLink:hover {
-        background-color: #fff;
-        color: #242424;
-        border-radius: 0;
-    }
-
-    .NavLogo {
-        color: #fff;
-        cursor: pointer;
-        text-decoration: none;
-        font-size: 2rem;
-        display: flex;
-        align-items: center;
-    }
-
-    .MenuIcon {
-        display: block;
-        position: absolute;
-        top: 0;
-        right: 0;
-        transform: translate(-100%, 60%);
-        font-size: 1.8rem;
-        cursor: pointer;
-    }
-
-    .fa-times {
-        color: #fff;
-        font-size: 2rem;
-    }
-}
