--- frontend/src/components/Time.module.css
@@ -1,95 +1,0 @@
-.TimeCards {
-    padding-right: 4rem;
-    padding-left: 4rem;
-    padding-top: 2rem;
-    padding-bottom: 2rem;
-}
-
-.TimeCardsItems {
-    text-align: center;
-    display: flex;
-    justify-content: space-between;
-    /* width: min-content; */
-}
-
-.TimeItem {
-    width: 350px;
-    display: inline-block;
-    /* border-radius: 10px; */
-}
-
-.TimeItemBlock {
-    height: 300px;
-    position: relative;
-    border-color: #fff;
-    border-style: solid;
-    border-width: 1px;
-    border-bottom: 0px;
-    border-radius: 20%;
-}
-
-.TimeItemFigure {
-    width: 200px;
-    height: 200px;
-    position: absolute;
-    top: 50%;
-    left: 50%;
-    transform: translate(-50%, -50%);
-}
-
-.TimeItemFigure h1 {
-    margin: 0;
-    position: absolute;
-    font-size: xxx-large;
-    top: 50%;
-    left: 50%;
-    transform: translate(-50%, -50%);
-}
-
-.TimeItemDate {
-    text-align: center;
-    font-size: large;
-    font-weight: lighter;
-}
-
-.TimeItemNar {
-    position: absolute;
-    bottom: -10%;
-    left: 50%;
-    transform: translate(-50%, 10%);
-    width: 300px;
-    height: 100px;
-    background-color: #0B2A77;
-    border-radius: 40%;
-}
-
-.TimeItemNar h2 {
-    font-size: x-large;
-    position: absolute;
-    bottom: 30%;
-    left: 50%;
-    transform: translate(-50%, -30%);
-}
-
-h2 {
-    font-weight: lighter;
-}
-
-.Monthly {
-    padding-top: 2rem;
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    margin-left: 25px;
-    margin-right: 25px;
-}
-
-.Month {
-    font-style: normal;
-    font-size: small;
-}
-
-.Year {
-    font-weight: lighter;
-    font-size: small;
-}
