+++ frontend/src/components/UserAuth/index.module.css
@@ -47,6 +47,6 @@
     color: aliceblue;
     padding: 1% 2% ;
     font-size: 15px;
-
+    border-radius: 3%;
 }
 /************************************************************************/
