--- frontend/src/components/header.tsx
@@ -1,6 +1,0 @@
-// experiment.test.ts
-import { Component} from 'react'; // literally anything, don't even have to use it
-
-it('test', () => {
-  console.log('test');
-});
