--- frontend/src/pages/AdminPage/Admin.tsx
@@ -1,12 +1,0 @@
-import React from 'react';
-import styles from '../../pages/Pages.module.css';
-
-function Admin() {
-    return (
-        <div className={styles.admin} >
-            <h1>Admin</h1>
-        </div>
-    )
-}
-
-export default Admin
