--- frontend/src/pages/Home.tsx
@@ -1,12 +1,0 @@
-import React from 'react';
-import styles from './Pages.module.css';
-
-function Home() {
-    return (
-        <div className={styles.home} >
-            <h1>Home</h1>
-        </div>
-    )
-}
-
-export default Home
