--- frontend/src/pages/Pages.module.css
@@ -1,49 +1,0 @@
-* {
-    box-sizing: border-box;
-    margin: 0;
-    padding: 0;
-    font-family: 'PT Sans', sans-serif;
-    /* background-color: #112542; */
-  }
-  
-  .home,
-  .users,
-  .admin {
-  /* display: flex;
-  height: 90vh;
-  align-items: center;
-  justify-content: center; */
-  /* font-size: 3rem; */
-  background-color: #07192c;
-  color: #fff;
-  }
-  
-  .home {
-    background-position: center;
-    background-size: cover;
-    background-repeat: no-repeat;
-    color: #fff;
-    font-size: 100px;
-  }
-  
-  .users {
-  /* background-image: url('./images/img-2.jpg'); */
-  background-position: center;
-  background-size: cover;
-  background-repeat: no-repeat;
-  color: #fff;
-  font-size: 100px;
-  }
-  
-  .admin {
-  /* background-image: url('./images/img-8.jpg'); */
-  background-position: center;
-  background-size: cover;
-  background-repeat: no-repeat;
-  color: #fff;
-  font-size: 100px;
-  }
-  
-  .header{
-    color: #fff;
-}
