--- frontend/src/pages/User.module.css
@@ -1,55 +1,0 @@
-.User {
-    color: #fff;
-}
-
-.UserPanels {
-    background-position: center;
-    background-size: cover;
-    background-repeat: no-repeat;
-    color: #fff;
-    font-size: 100px;
-}
-
-.UserMain {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    margin-left: 25px;
-    margin-right: 25px;
-}
-
-.UserLeft {
-    display: inline-block;
-}
-
-.fa-user-circle {
-    font-size: 100px;
-}
-
-.UserLeftRight {
-    float: right;
-}
-
-.UserMiddle {
-    display: inline-block;
-}
-
-.UserRight {
-    display: inline-block;
-}
-
-.Demo {
-    padding-top: 2rem;
-}
-
-.Demo h1 {
-    padding-left: 2rem;
-}
-
-.Member {
-    padding-top: 2rem;
-}
-
-.Member h1 {
-    padding-left: 2rem;
-}
