--- frontend/src/pages/regular_users/Narratives.tsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-function Narratives() {
-    <h1>Narratives</h1>
-}
-
-export default Narratives();
