--- frontend/src/pages/regular_users/TimeView.tsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-function TimeView() {
-    <h1>Time View</h1>
-}
-
-export default TimeView();
