--- frontend/src/pages/regular_users/UsersSettings.tsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-function UserSettings() {
-    <h1>Settings</h1>
-}
-
-export default UserSettings();
