--- frontend/src/pages/regular_users/ViewSchedule.tsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-function ViewSchedule() {
-    <h1>Schedule</h1>
-}
-
-export default ViewSchedule();
