+++ frontend/src/styles/App.css
@@ -1,11 +1,10 @@
 
 .App {
-  text-align: left;
-  background-color: #07192c;
+  background-color: transparent;
 }
 
 .App-header {
-  min-height: 60px;
+  
   display: flex;
   flex-direction: column;
   align-items: center;
